requirements.txt
Flask==2.3.3
requests==2.31.0
fastapi==0.100.0
uvicorn==0.23.1
sqlalchemy==2.1.0
bcrypt==4.0.1
python-jose==3.3.0
starlette==0.27.0
pydantic==1.11.1
